+++ packages/nhost-js/CHANGELOG.md
@@ -1,5 +1,6 @@
 # dev
 
+- feat (packages/nhost-js): graphl client's post now supports TypedDocumentNode #31
 - fix (tools/codegen;packages/nhost-js): encode URL parameters better #30
 - fix (packages/next-js): use proper types for Webauthn endpoints #29
 - chore (ci): enabled #26